added useEditingCells.ts
